+++ src/components/MyServiceList/index.tsx
@@ -5,19 +5,19 @@ import * as S from './style';
 import EmptyList from './EmptyList';
 import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
-import { UserLists } from '../../Atom/Atoms';
+import { User } from '../../Atom/Atoms';
 
 export default function MyServiceList() {
   const [modifyItem, setModifyItem] = useState<ClientListType | null>(null);
-  const userList = useRecoilValue(UserLists);
+  const user = useRecoilValue(User);
 
   return (
     <S.Layer>
       <S.Title>내가 등록한 서비스</S.Title>
-      {userList.length !== 0 ? (
+      {user.clientList.length !== 0 ? (
         <S.ListWrapper>
           <>
-            {userList.map((listItem, index) => {
+            {user.clientList.map((listItem, index) => {
               return <ListItem key={index} listData={listItem} />;
             })}
           </>
